requirements.txt
colorama
requests
python-dateutil